separated parsing part of Request.cpp into Request_parsing.cpp
